+++ src/constants/loggingMethods.js
@@ -1,1 +1,3 @@
+/* @flow */
+
 export default ['log', 'trace', 'debug', 'info', 'warn', 'error'];