+++ src/globalFlowTypes.js
@@ -1,0 +1,13 @@
+import Clogy from './main/Clogy';
+
+export type LevelsType = {
+  log: number,
+  trace: number,
+  debug: number,
+  info: number,
+  warn: number,
+  error: number,
+  none: number
+};
+
+export type ClogyType = Clogy;