+++ src/index.js
@@ -1,6 +1,10 @@
+/* @flow */
+
 import Clogy from './main/Clogy';
 
-const clogy = new Clogy();
+import type { ClogyType } from './globalFlowTypes';
+
+const clogy: ClogyType = new Clogy();
 
 export default clogy;
 
@@ -8,4 +12,6 @@ export default clogy;
 // Can use plugin: https://www.npmjs.com/package/babel-plugin-add-module-exports
 // Used this soln. instead:
 // http://stackoverflow.com/questions/34736771/webpack-umd-library-return-object-default/34778391#34778391
+//
+// $FlowFixMe: suppressing this error until babel changes this style
 module.exports = clogy;