+++ src/main/Clogy.js
@@ -1,8 +1,19 @@
+/* @flow */
+
 import singleton from '../utilities/singleton';
 import Logger from './Logger';
 
 class Clogy extends Logger {
-  noConflict() {
+
+  /**
+   * If you are using another JavaScript library that uses the clogy variable,
+   * you can run into conflicts with this library. In order to avoid these
+   * conflicts, you need to put clogy in no-conflict mode immediately after it
+   * is loaded onto the page and before you attempt to use clogy in your page.
+   * It works similar to jQuery's no-conflict mode
+   * @return {clogy} Returns current instance i.e. clogy
+   */
+  noConflict(): this {
 
     // No need of ===, typeof returns a string
     if (typeof(window) != 'undefined' && window.clogy) {
@@ -16,7 +27,17 @@ class Clogy extends Logger {
     return this;
   }
 
-  decorator(decoFunc) {
+  /**
+   * Used to extend logging functionality. Can be used for:
+   * 1. Adding a prefix
+   * 2. Submitting logs to server
+   * 3. Logging to a file
+   * 4. Showing toast messages
+   * @param  {Function} decoFunc: A decorator callback for extending logging
+   *                               functionality
+   * @return {void | undefined} Returns nothing
+   */
+  decorator(decoFunc: () => void): void {
 
     // No need of ===, typeof returns a string
     if (typeof(decoFunc) != 'function') {