+++ src/utilities/logging.js
@@ -1,6 +1,8 @@
+/* @flow */
+
 import LOGGING_LEVELS from '../constants/loggingLevels';
 
-const DEFAULT_LOGGING_TYPE = 'log';
+const DEFAULT_LOGGING_TYPE: string = 'log';
 
 // I prefer default export rather than named export
 export default {
@@ -9,13 +11,23 @@ export default {
 
 ////////////////////////
 
-function isConsoleDefined() {
+/**
+ * Check if console is available or not
+ * @return {Boolean} Returns true or false
+ */
+function isConsoleDefined(): boolean {
 
   // No need of ===, typeof returns a string
   return typeof(console) != 'undefined';
 }
 
-function isLogLevelValid(currentLogLevel) {
+/**
+ * Check if current log level is valid, less than max
+ * and more than min log level
+ * @param  {Number}  currentLogLevel: Current log level eg. 1, 2, 3, 4
+ * @return {Boolean} Returns true or false
+ */
+function isLogLevelValid(currentLogLevel: number): boolean {
   return !!(
     currentLogLevel &&
     currentLogLevel >= LOGGING_LEVELS.range.min &&
@@ -23,53 +35,114 @@ function isLogLevelValid(currentLogLevel) {
   );
 }
 
-function isNoneLogLevel(currentLogLevel) {
+/**
+ * Check if current log level is none thus logging is disabled
+ * @param  {Number}  currentLogLevel: Current log level eg. 1, 2, 3, 4
+ * @return {Boolean} Returns true or false
+ */
+function isNoneLogLevel(currentLogLevel: number): boolean {
   return currentLogLevel === LOGGING_LEVELS.types.none;
 }
 
-function isLoggingAllowed(currentLogLevel, loggingType) {
+/**
+ * Check if current log level is more than level of current log method thus
+ * allowing logging accordingly
+ * @param  {Number}  currentLogLevel: Current log level eg. 1, 2, 3, 4
+ * @param  {String}  loggingType    : logging method eg. log, info, error, warn
+ * @return {Boolean} Returns true or false
+ */
+function isLoggingAllowed(
+  currentLogLevel: number,
+  loggingType: string
+): boolean {
   return (
     LOGGING_LEVELS.types[loggingType] >= currentLogLevel
   );
 }
 
-function getDateTime() {
-  const d = new Date();
-  const date = d.toDateString();
+/**
+ * Get current date and time
+ * @return {String} Returns date and time in string format
+ */
+function getDateTime(): string {
+  const d: Object = new Date();
+  const date: string = d.toDateString();
 
   // 0 means first matched string, which will be true always unless someone
   // overrides the result
-  const time = d.toTimeString().match(/^([0-9]{2}:[0-9]{2}:[0-9]{2})/)[0];
+  const time: string = d.toTimeString().match(/^([0-9]{2}:[0-9]{2}:[0-9]{2})/)[0];
 
   // NOTE: hey..!! see, 3 is not a magic number, if you think it is, then
   // it's just milliseconds in 3 digits (000-999)
-  const appendMilliseconds = `00${d.getMilliseconds()}`.slice(-3);
+  const appendMilliseconds: string = `00${d.getMilliseconds()}`.slice(-3);
 
   return `${date} ${time}.${appendMilliseconds}`;
 }
 
-function justLogItDude(loggingType, args) {
-  const consoleLoggingType = console[loggingType] || console[DEFAULT_LOGGING_TYPE];
+/**
+ * Get console options and push them into an array (for logging)
+ * @param  {Object} options: Console options like showDateTime, prefix
+ * @return {any[]} Returns an array of console options
+ */
+function getConsoleOptions(options: Object): any[] {
+  const consoleOptions: any[] = [];
 
-  // No need of ===, typeof returns a string
-  if (typeof(consoleLoggingType) != 'function') {
+  if (options.showDateTime) {
 
-    // I don't want to break an application if someone intentionally
-    // deleted the logging methods, so skipping error here
+    // Better readability than consoleOptions[consoleOptions.length] = value
+    consoleOptions.push(`${getDateTime()}: `);
+  }
+
+  if (options.prefix) {
+    consoleOptions.push(options.prefix);
+  }
+
+  return consoleOptions;
+}
+
+/**
+ * Used for logging to console
+ * @param  {String} loggingType: logging method eg. log, info, error, warn
+ * @param  {any}    args       : any value
+ * @return {void | undefined} Returns nothing
+ */
+function justLogItDude(loggingType: string, args: any[]): void {
+  const consoleLoggingType: () => void = console[loggingType] || console[DEFAULT_LOGGING_TYPE];
+
+  // I know this should be returned earlier but there is
+  // negligible performance hit.
+  // May be I can move this to a function which can return me
+  // consoleLoggingType or false (maintain S of SOLID),
+  // but I prefer this to be here
+  if (!consoleLoggingType) {
     return;
   }
 
   Function.prototype.bind.call(consoleLoggingType, console).apply(console, args);
 }
 
-// Functional style programming;
-// No mutating params, no state known beforehand
-function logToConsole(logToConsoleParams, args) {
+/**
+ * Used for logging to console
+ * Functional style programming; No mutating params, no state known beforehand
+ * @param  {Object} logToConsoleParams: Object containing current log level ,
+ *                                      log methods, default options like prefix
+ * @param  {any}    args              : any value
+ * @return {void | undefined} Returns nothing
+ */
+function logToConsole(logToConsoleParams: Object, args: any[]): void {
   if (!isConsoleDefined()) {
     return;
   }
 
-  const { currentLogLevel, loggingType, options = {} } = logToConsoleParams;
+  const {
+    currentLogLevel,
+    loggingType,
+    options = {}
+  }: {
+    currentLogLevel: number,
+    loggingType: string,
+    options: Object
+  } = logToConsoleParams;
 
   if (!isLogLevelValid(currentLogLevel)) {
     throw new RangeError('Invalid LogLevel set, Please set a valid LogLevel');
@@ -82,18 +155,5 @@ function logToConsole(logToConsoleParams, args) {
     return;
   }
 
-  /////////////////////////////////////////////
-  const consoleOptions = [];
-
-  if (options.showDateTime) {
-
-    // Better readability than consoleOptions[consoleOptions.length] = value
-    consoleOptions.push(`${getDateTime()}: `);
-  }
-
-  if (options.prefix) {
-    consoleOptions.push(options.prefix);
-  }
-
-  justLogItDude(loggingType, [...consoleOptions, ...args]);
+  justLogItDude(loggingType, [...getConsoleOptions(options), ...args]);
 }